test: address review comments on read_contract bytes normalization
- Assert the wrapped tool invocation and default arguments in the
  read_contract bytes-result route test (per API route-test guideline).
- Split the _wrap_tool_for_structured_output tests into a focused module
  tests/test_server_structured_output.py so test_server.py stays under the
  500-LOC limit; register the new module in AGENTS.md.
